% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/approx_fs.R
\name{approximate_fs}
\alias{approximate_fs}
\title{Approximate forward stepwise selection}
\usage{
approximate_fs(x, y, k_max = 10)
}
\arguments{
\item{x}{The input matrix.  Assumed to be the logarithmic transform of positive raw inputs.}

\item{y}{The real-valued response.}

\item{k_max}{The largest number of log-ratios to consider.}
}
\value{
theta The k_max by k_max upper triangular matrix of coefficient for each log-ratio at each step.

ratios A 2 by k_max grid giving the selected ratios

beta A p by k_max matrix where column j gives the coefficient value after j steps.

intercept The intercept for each model.
}
\description{
Fits the approximate forward stepwise procedure.
}
