% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constrained_lasso.R
\name{cv.glmnet.constr}
\alias{cv.glmnet.constr}
\title{CV for constrained lasso.}
\usage{
cv.glmnet.constr(fit, x, y, nfolds = 10, foldid = NULL, keep = F,
  alpha = 1)
}
\arguments{
\item{fit}{The result of a "glmnet.contsr" function call.}

\item{x}{The matrix of features.}

\item{y}{The response value.  In the real-valued repsonse case, y should be centered.}

\item{nfolds}{The number of CV folds. Default is 10.}

\item{foldid}{Fold id for each observation.  By default these are uniformly random.}

\item{alpha}{Elastic net parameter.  Default is 1 which gives the lasso.}
}
\value{
lambda The gride of lambda values

cvm The CV estimate of prediction error for each lambda.

yhat.preval The predicted values generated by each fold.
}
\description{
Cross-validation to choose lambda in the constrained lasso procedure
}
