% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/approx_fs.R
\name{cv_approximate_fs}
\alias{cv_approximate_fs}
\title{CV for approximate forward stepwise selection}
\usage{
cv_approximate_fs(x, y, k_max = 5, n_folds = 10, fold_id = NULL, ...)
}
\arguments{
\item{x}{The input matrix.  Assumed to be the logarithmic transform of positive raw inputs.}

\item{y}{The real-valued response.}

\item{k_max}{The largest number of log-ratios to consider.}

\item{n_folds}{The number of folds.  Default is 10.}

\item{fold_id}{Optional fold assignments.  Default is uniformly random assignment.}
}
\value{
cvm The CV estimate of prediction error for each model size.

beta The coefficient vector corresponding to the best model size chosen by CV.

intercept.  The intercept of the best model chosen by CV.

model A approximate_fs object with k_max set to by the optimal model size chose by CV.
}
\description{
Determines the number of steps of approximate forward stepwise selection by cross-validation.
}
