% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/two_stage.R
\name{cv_two_stage}
\alias{cv_two_stage}
\title{CV for two-stage}
\usage{
cv_two_stage(z, y, lambda_1 = NULL, k_max = 5, n_folds = 10, ...)
}
\arguments{
\item{z}{The input matrix.  Assumed to be the logarithmic transform of positive raw inputs.}

\item{y}{The real-valued response.}

\item{lambda_1}{Optional vector of lambda values.}

\item{k_max}{The largest number of log-ratios to consider.}

\item{n_folds}{Optional number of folds.  Default is 10.}

\item{...}{Additional arguments passed to "two_stage"}
}
\value{
mse A matrix of MSE values generated by CV.

best_params The best lambda and k parameter values chosen by CV.

lambda_min A best value of lambda.

k_min A best value of k.

two_step_obj The two_step object fit on the full data.

beta_min The parameter value selected by CV.

lambda The grid of lambda values used.
}
\description{
Cross-validation for the lambda and k parameters in the two-stage log-ratio lasso.
}
