% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constrained_lasso.R
\name{glmnet.constr}
\alias{glmnet.constr}
\title{Constrained lasso fitting.}
\usage{
glmnet.constr(x, y, family = c("gaussian", "binomial"), alpha = 1,
  nlambda = 100, lambda.min.ratio = 0.01, lambda = NULL, ...)
}
\arguments{
\item{x}{The matrix of features.}

\item{y}{The response value.  In the real-valued repsonse case, y should be centered.}

\item{family}{The model family. Default is "gaussian" for standard lasso.
"binomial" is available for lasso logistic regression.}

\item{alpha}{The elastic net parameter.  Default is 1 (the standard lasso).}

\item{nlambda}{Numer of lambda values to fit}

\item{lambda.min.ratio}{Smallest value for lambda as a fraction of lambda.max}

\item{lambda}{Optional grid of lambda values}

\item{...}{Additional arguments to internally pass to the "glmnet" function.}
}
\value{
beta The resulting coefficient vectors in matrix form.
}
\description{
Run the constrained lasso procedure.
}
