% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/two_stage.R
\name{two_stage}
\alias{two_stage}
\title{Two-stage log-ratio lasso}
\usage{
two_stage(z, y, k_max = 20, lambda_1 = NULL, second.stage = c("y",
  "yhat"), ...)
}
\arguments{
\item{z}{The input matrix.  Assumed to be the logarithmic transform of positive raw inputs.}

\item{y}{The real-valued response.}

\item{k_max}{The largest number of log-ratios to consider.}

\item{lambda_1}{Optional vector of lambda values.}

\item{second.stage}{Specifies whether to fit the second stage on the response y or the fitted values yhat.
Fitting on yhat is known as "conservative two-stage" in the log-ratio lasso paper.}

\item{...}{Additional arguments passed to "glmnet.constr".}
}
\value{
beta A list of p by k_max matrices where column j gives the coefficient value after j steps.
  each list entry corresponds to a value of lambda.

coef. The theta coefficients selected for each model.  Recommended for internal use only.

lambda The grid of lambda values used for the fitting.

selected_vars A matrix showing which log-terms are active at each step of the second-stage
forward stepwise procedure.
}
\description{
Fits the two-stage log-ratio lasso.
}
